refactor(frontend): group listing page shells under pages/listing/
Move the taxonomy listing page shells (TaxonomyListPage, GroupedTaxonomyList)
and the shared NoResultsCreatePrompt into pages/listing/, beside the SSR
listing controllers already there. All three are multi-route page bodies;
NoResultsCreatePrompt's only non-route consumers are listing shells now
co-located with it.

Add pages/README.md describing the page-shell convention and the routes/-or-
within-pages/ import boundary.
