+++ frontend/src/lib/components/pages/listing/FacetedCatalogListing.svelte
@@ -10,7 +10,7 @@
     type FilterChipSpec,
   } from '$lib/components/collections/filters/ActiveFilterChips.svelte';
   import FilterDrawer from '$lib/components/collections/filters/FilterDrawer.svelte';
-  import NoResultsCreatePrompt from '$lib/components/NoResultsCreatePrompt.svelte';
+  import NoResultsCreatePrompt from './NoResultsCreatePrompt.svelte';
   import SearchBox from '$lib/components/ui/SearchBox.svelte';
   import PaginatedListLoader from './PaginatedListLoader.svelte';
   import { ENTITY_META, type CatalogEntityKey } from '$lib/entities/entity-meta';
