+++ frontend/src/lib/components/pages/listing/GroupedTaxonomyList.svelte
